pyproject.toml
[tool.poetry]
name = "sigmabackend"
version = "0.1.0"
description = ""
authors = ["Henish Patel <henishpatel9045@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
cloudinary = "^1.41.0"
python-decouple = "^3.8"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
